assert-preparation/generator/pyproject.toml
[project]
name = "generator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "cairosvg>=2.9.0",
    "google-genai>=1.73.1",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
]
